fix: change from node-fetch to await import again
Node-fetch just doesnt work if we want it to run on every different
environment. In the previous commit, everything worked fine in tests,
development and preview (production) mode but it failed when the tests
run in the github runner environement because of the constructed
absolute path that is needed for node-fetch. Thus I revert everything
back to await import when getData is used in node environement and using
ordinary fetch in browser environment.