feat: reduce complexity in visibility explanations
Previously we had three explanations per visiblity (6 explanations per
data type) which is hard to maintain. Now, I changed to so that we have
two explanations for the visibility per data type). We can still refer
to the current role with attributes and the risk.json aren't so bloated
anymore. This is currently only implemented for test data, real data
will be implemented later.